% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain_state.R
\name{chain_state}
\alias{chain_state}
\title{Construct a new chain state.}
\usage{
chain_state(position, momentum = NULL)
}
\arguments{
\item{position}{Position component of chain state.}

\item{momentum}{Momentum component of chain state. Optional.}
}
\value{
New chain state object.
}
\description{
The chain state object provides cached access to target distribution
log density and its gradient.
}
\examples{
state <- chain_state(c(0.1, -0.5))
target_distribution <- list(
  log_density = function(x) sum(x^2) / 2,
  grad_log_density = function(x) x
)
state$grad_log_density(target_distribution)
}
