% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampling.R
\name{gibbs}
\alias{gibbs}
\title{Gibbs sampling implementation}
\usage{
gibbs(cond, init, n)
}
\arguments{
\item{cond}{list of conditional distributions}

\item{init}{vector of inital values of the chains in}

\item{n}{sample size}
}
\value{
a \code{dataframe} with the given list keys from \code{cond} as variables
}
\description{
Gibbs sampling implementation
}
\examples{
 # instantiating conditionals
 cond <- list(x_1 = function(x_2){return(rnorm(1, x_2))},
              x_2 = function(x_1){return(rnorm(1, x_1))})

 # simualte sample
 sample <- gibbs(cond, c(1,1), 1000)

 # plot sample
 hist(sample$x_1)
 hist(sample$x_2)
}
