% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamiltonian.R
\name{involution_hamiltonian}
\alias{involution_hamiltonian}
\title{Apply involution underlying Hamiltonian proposal to a chain state.}
\usage{
involution_hamiltonian(state, n_step, scale_and_shape, target_distribution)
}
\value{
Chain state after involution is applied.
}
\description{
Apply involution underlying Hamiltonian proposal to a chain state.
}
\keyword{internal}
