% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langevin.R
\name{involution_langevin}
\alias{involution_langevin}
\title{Apply involution underlying Langevin proposal to a chain state.}
\usage{
involution_langevin(state, scale_and_shape, target_distribution)
}
\value{
Chain state after involution is applied.
}
\description{
Apply involution underlying Langevin proposal to a chain state.
}
\keyword{internal}
