% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log1p_exp}
\alias{log1p_exp}
\title{Numerically stable computation of \code{log(1 + exp(x))}.}
\usage{
log1p_exp(x)
}
\arguments{
\item{x}{Scalar or vector to evaluate function at.}
}
\value{
Value of \code{log(1 + exp(x))}
}
\description{
Numerically stable computation of \code{log(1 + exp(x))}.
}
