% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barker.R
\name{log_density_ratio_barker}
\alias{log_density_ratio_barker}
\title{Compute logarithm of Barker proposal density ratio.}
\usage{
log_density_ratio_barker(state, proposed_state, target_distribution)
}
\arguments{
\item{state}{Current chain state.}

\item{proposed_state}{Proposed chain state.}

\item{target_distribution}{Target stationary distribution for chain.}
}
\value{
Logarithm of proposal density ratio.
}
\description{
Compute logarithm of Barker proposal density ratio.
}
\examples{
state <- chain_state(c(0., 0.))
target_distribution <- list(
  log_density = function(x) sum(x^2) / 2,
  grad_log_density = function(x) x
)
proposed_state <- sample_barker(state, target_distribution, step_size = 1.)
log_density_ratio_barker(
  state, proposed_state, target_distribution
)
}
