% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logistic_sigmoid}
\alias{logistic_sigmoid}
\title{Logistic sigmoid function \code{1 / (1 + exp(-x))}.}
\usage{
logistic_sigmoid(x)
}
\arguments{
\item{x}{Scalar or vector evaluate logistic sigmoid at.}
}
\value{
Value of logistic sigmoid at \code{x}.
}
\description{
Logistic sigmoid function \code{1 / (1 + exp(-x))}.
}
