% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{min_1_exp}
\alias{min_1_exp}
\title{Numerically stable computation of \code{min(1, exp(x))}.}
\usage{
min_1_exp(x)
}
\arguments{
\item{x}{Scalar or vector to evaluate function at.}
}
\value{
Value of \code{min(1, exp(x))}
}
\description{
Numerically stable computation of \code{min(1, exp(x))}.
}
