% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_normal_sampling.R
\name{MVNormGibbs}
\alias{MVNormGibbs}
\title{Gibbs MVnormal sampling implementation}
\usage{
MVNormGibbs(n, mu, simga, init)
}
\arguments{
\item{n}{integer for size of sample}

\item{mu}{vector for mean values}

\item{init}{vector of inital values of the chains in}

\item{sigma}{matrix for covariance}
}
\value{
a \code{dataframe} as the sample
}
\description{
Gibbs MVnormal sampling implementation
}
