% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneDSliceDebug.R
\name{OneDSliceDebug}
\alias{OneDSliceDebug}
\title{One D slice implementation with intervalls}
\usage{
OneDSliceDebug(it, x0, w, f, L.lim = -Inf, R.lim = Inf)
}
\arguments{
\item{it}{integer: number of iterations (or sample size)}

\item{x0}{numeric: starting value}

\item{w}{numeric: width of initial intervall}

\item{f}{function: density to be modeled}

\item{L.lim}{numeric: lower limit for candidate slice (default \code{-Inf})}

\item{R.lim}{numeric: upper limit for candidate slice (default \code{Inf})}
}
\description{
This implementation does not need a function to determine the reverse image. Instead holistic techniques will be used, similar to the accept reject condition.
This implementation is a debug version of the original. Additionally to the sample it returns the interval bounds.
}
\examples{
## Normal distribution:
 xnormal <- OneDSlice(10000, x0 = 1, w = 1/8,
                      f = function(x) {exp(-x^2/2)/sqrt(2 * pi)})

 hist(xnormal)
 ## Works.

 ## Multimodal case:
 xmm <- OneDSlice(10000, x0 = 1, w = 1/8,
                  f = function(x) {exp(-x^2/2)/sqrt(2 * pi) +
                      exp(-(x - 4)^2/2)/sqrt(2 * pi)})

 hist(xmm, freq = FALSE, ylim = c(0, 0.25), breaks = 100)
 curve((exp(-x^2/2)/sqrt(2 * pi) +
          exp(-(x - 4)^2/2)/sqrt(2 * pi))/2,
       from = -4, to = 8, add = TRUE)
 ## works as well.

 xmm2 <- OneDSlice(10000, x0 = 1, w = 1/8,
                   f = function(x) {exp(-x^2/2)/sqrt(2 * pi) +
                       exp(-(x - 4)^2/2)/sqrt(2 * pi) +
                       exp(-((x + 5)/3)^2/2)/sqrt(2 * pi * 9)})

 hist(xmm2, freq = FALSE, ylim = c(0, 0.15), breaks = 100)
 curve((exp(-x^2/2)/sqrt(2 * pi) +
          exp(-(x - 4)^2/2)/sqrt(2 * pi)+
          exp(-((x + 5)/3)^2/2)/sqrt(2 * pi * 9))/3,
       from = -15, to = 8, add = TRUE)
 ## Looks perfect.

 ## Try out case with bad initial value:
 ##...

 ## What if distribution is bounded, like exponential distribution?
 ## In this case we need to set the lower limit for the slice L.lim = 0.

 xexp <- OneDSlice(10000, x0 = 1, w = 1/8,
                   f = function(x) {0.5 * exp(-x * 0.5)},
                   L.lim = 0)

 hist(xexp, freq = FALSE, breaks = 100)
 curve(0.5 * exp(-x * 0.5), from = 0, to = 10, add = TRUE)



}
