% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baysian_logit_regression_slice.R
\name{p}
\alias{p}
\title{Utility for bayesian logistic regression}
\usage{
p(beta, y, X, prior, scaling = 1)
}
\description{
This function implements a kernel of a posterior distribution. Do not use this function. Use \code{bayes_logit_reg_slice} instead.
}
