% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_labels_bayes.R
\name{predict_labels_bayes_logit}
\alias{predict_labels_bayes_logit}
\title{predicting labels for bayesian logistic regression}
\usage{
predict_labels_bayes_logit(beta, X, threshold = 0.5)
}
\arguments{
\item{beta}{vector: some beta parameters}

\item{X}{martix: new regression matrix}

\item{threshold}{numeric: Threshold for switching to \eqn{\hat y = 1}}
}
\description{
predicting labels for bayesian logistic regression
}
\examples{
# example code
 n <- 1000

 # simulating sample
 X <- matrix(data = c(rep(1, n), seq(-10, 10, length.out = n)), ncol=2)
 props <- expit(X[,2]+2)
 sample_y <- function(prob) {as.numeric(runif(1) <= prob)}
 y <- sapply(props, sample_y)

 beta_samples <- bayesian.logit.slice(1000, y, X, scaling = 1000)

 beta_hat <- apply(beta_samples, 2, mean)
 X_test <- matrix(data = c(rep(1, 100), seq(-10, 10, length.out = 100)), ncol=2)

 predict_labels_bayes_logit(beta_hat, X_test)

}
