% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barker.R
\name{sample_barker}
\alias{sample_barker}
\title{Sample new state from Barker proposal.}
\usage{
sample_barker(x, grad_x, raux = stats::rnorm, runif = stats::runif, ...)
}
\arguments{
\item{x}{Current chain state.}

\item{grad_x}{Gradient of target distribution log density function at current
chain state.}

\item{raux}{Function which generates a random vector from auxiliary variable
distribution.}

\item{runif}{Function which generates a random vector from standard uniform
distribution given an integer size.}

\item{...}{Any additional arguments to pass to \code{raux} function.}
}
\value{
Proposed new state.
}
\description{
Sample new state from Barker proposal.
}
\examples{
x <- c(0., 0.)
grad_x <- c(-0.1, 0.5)
sample_barker(x, grad_x, sd = 2.)
}
