% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamiltonian.R
\name{sample_hamiltonian}
\alias{sample_hamiltonian}
\title{Sample new state from Hamiltonian proposal.}
\usage{
sample_hamiltonian(
  state,
  target_distribution,
  n_step,
  scale_and_shape,
  sample_auxiliary = stats::rnorm
)
}
\description{
Sample new state from Hamiltonian proposal.
}
\keyword{internal}
