% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{sample_metropolis_hastings}
\alias{sample_metropolis_hastings}
\title{Sample from Metropolis-Hastings kernel.}
\usage{
sample_metropolis_hastings(
  state,
  target_distribution,
  proposal,
  sample_uniform = stats::runif
)
}
\arguments{
\item{state}{Current chain state.}

\item{target_distribution}{Target distribution for chain.}

\item{proposal}{Proposal distribution.}

\item{sample_uniform}{Function which generates a random vector from standard
uniform distribution given an integer size.}
}
\value{
List with named entries \code{state} corresponding to new chain state and
\code{statistics}, a list with named entries for statistics of transition, here
this consisting of a named entry \code{accept_prob} for the Metropolis
acceptance probability.
}
\description{
Sample from Metropolis-Hastings kernel.
}
\examples{
target_distribution <- list(
  log_density = function(x) sum(x^2) / 2,
  grad_log_density = function(x) x
)
proposal <- barker_proposal(target_distribution, step_size = 1.)
n_sample <- 1000
states <- vector("list", n_sample)
states[[1]] <- chain_state(rnorm(2))
for (s in 2:n_sample) {
  state_and_statistics <- sample_metropolis_hastings(
    states[[s - 1]], target_distribution, proposal
  )
  states[[s]] <- state_and_statistics$state
}
}
