% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridges.R
\name{trace_function_from_stan_model}
\alias{trace_function_from_stan_model}
\title{Construct trace function from a BridgeStan \code{StanModel} object.}
\usage{
trace_function_from_stan_model(
  model,
  include_log_density = TRUE,
  include_generated_quantities = FALSE,
  include_transformed_parameters = FALSE
)
}
\arguments{
\item{model}{Stan model object to use to generate (constrained) parameters to
trace.}

\item{include_log_density}{Whether to include an entry \code{log_density}
corresponding to current log density for target distribution in values
returned by trace function.}

\item{include_generated_quantities}{Whether to included generated quantities
in Stan model definition in values returned by trace function.}

\item{include_transformed_parameters}{Whether to include transformed
parameters in Stan model definition in values returned by trace function.}
}
\value{
A function which given \code{chain_state} object returns a named vector of
values to trace during sampling. The constrained parameter values of model
will always be included.
}
\description{
Construct trace function from a BridgeStan \code{StanModel} object.
}
\examples{
\dontshow{if (requireNamespace("bridgestan", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
model <- example_gaussian_stan_model()
trace_function <- trace_function_from_stan_model(model)
withr::with_seed(876287L, state <- chain_state(rnorm(model$param_unc_num())))
trace_function(state)
\dontshow{\}) # examplesIf}
}
