% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptation.R
\name{variance_adapter}
\alias{variance_adapter}
\title{Create object to adapt proposal with per dimension scales based on estimates
of target distribution variances.}
\usage{
variance_adapter(kappa = 0.6)
}
\arguments{
\item{kappa}{Decay rate exponent in \verb{[0.5, 1]} for adaptation learning rate.}
}
\value{
List of functions with entries
\itemize{
\item \code{initialize}, a function for initializing adapter state and proposal
parameters at beginning of chain,
\item \code{update} a function for updating adapter state and proposal parameters on
each chain iteration,
\item \code{finalize} a function for performing any final updates to adapter state and
proposal parameters on completion of chain sampling (may be \code{NULL} if
unused).
\item \code{state} a zero-argument function for accessing current values of adapter
state variables.
}
}
\description{
Create object to adapt proposal with per dimension scales based on estimates
of target distribution variances.
}
\examples{
target_distribution <- list(
  log_density = function(x) -sum(x^2) / 2,
  grad_log_density = function(x) -x
)
proposal <- barker_proposal(target_distribution)
adapter <- variance_adapter()
adapter$initialize(proposal, chain_state(c(0, 0)))
}
