% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barker.R
\name{log_density_ratio_barker}
\alias{log_density_ratio_barker}
\title{Compute logarithm of Barker proposal density ratio.}
\usage{
log_density_ratio_barker(x, y, grad_x, grad_y)
}
\arguments{
\item{x}{Current chain state.}

\item{y}{Proposed chain state.}

\item{grad_x}{Gradient of target distribution log density function at current
chain state.}

\item{grad_y}{Gradient of target distribution log density function at
proposed chain state.}
}
\value{
Logarithm of proposal density ratio.
}
\description{
Compute logarithm of Barker proposal density ratio.
}
\examples{
x <- c(0., 0.)
grad_x <- c(-0.1, 0.5)
y <- c(1., 0.)
grad_y <- c(0.7, -0.4)
log_density_ratio_barker(x, y, grad_x, grad_y)
}
