% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barker.R
\name{sample_barker}
\alias{sample_barker}
\title{Sample new state from Barker proposal.}
\usage{
sample_barker(
  state,
  target_distribution,
  scale_and_shape,
  sample_auxiliary = stats::rnorm,
  sample_uniform = stats::runif
)
}
\arguments{
\item{state}{Current chain state.}

\item{target_distribution}{Target stationary distribution for chain. A list
with named entries \code{log_density} and \code{grad_log_density} corresponding to
respectively functions for evaluating the logarithm of the (potentially
unnormalized) density of the target distribution and its gradient.
As an alternative to \code{grad_log_density} an entry
\code{value_and_grad_log_density} may instead be provided which is a function
returning both the value and gradient of the logarithm of the (unnormalized)
density of the target distribution as a list under the names \code{value} and
\code{grad} respectively.}

\item{scale_and_shape}{Scalar, vector or matrix which scales and shapes
proposal distribution. If a scalar (in which case the value should be
non-negative) the auxiliary vector will be isotropically scaled by the
value. If a vector (in which case the value should be equal in length to the
dimension of the space and all entries non-negative) each dimension of the
auxiliary vector will be scaled separately. If a matrix (in which case the
value should be a square matrix with size equal to the dimension of the
space) then by pre-multiplying the auxiliary vector arbitrary linear
transformations can be performed.}

\item{sample_auxiliary}{Function which generates a random vector from
auxiliary variable distribution.}

\item{sample_uniform}{Function which generates a random vector from standard
uniform distribution given an integer size.}
}
\value{
Proposed new chain state.
}
\description{
Sample new state from Barker proposal.
}
\keyword{internal}
