% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chains.R
\name{sample_chain}
\alias{sample_chain}
\title{Sample a Markov chain}
\usage{
sample_chain(
  target_distribution,
  proposal,
  initial_state,
  n_warm_up_iteration,
  n_main_iteration,
  adapters = NULL,
  trace_function = NULL,
  seed = NULL,
  show_progress_bar = TRUE
)
}
\arguments{
\item{target_distribution}{Target stationary distribution for chain. A list
with named entries \code{log_density} and \code{grad_log_density} corresponding to
respectively functions for evaluating the logarithm of the (potentially
unnormalized) density of the target distribution and its gradient.
As an alternative to \code{grad_log_density} an entry
\code{value_and_grad_log_density} may instead be provided which is a function
returning both the value and gradient of the logarithm of the (unnormalized)
density of the target distribution as a list under the names \code{value} and
\code{grad} respectively.}

\item{proposal}{Proposal distribution object. Must define entries \code{sample}, a
function to generate sample from proposal distribution given current chain
state and \code{log_density_ratio}, a function to compute log density ratio for
proposal for a given pair of current and proposed chain states.}

\item{initial_state}{Initial chain state.}

\item{n_warm_up_iteration}{Number of warm-up (adaptive) chain iterations to
run.}

\item{n_main_iteration}{Number of main (non-adaptive) chain iterations to
run.}

\item{adapters}{List of adapters to tune proposal parameters during warm-up.}

\item{trace_function}{Function which given current chain state outputs list
of variables to trace on each main (non-adaptive) chain iteration.}

\item{seed}{Integer value to seed pseudo-random number generator with.}

\item{show_progress_bar}{Whether to show progress bars during sampling.
Requires \code{progress} package to be installed to have an effect.}
}
\value{
A list with entries
\itemize{
\item \code{final_state}: the final chain state,
\item \code{traces}: a dataframe contained traced variables for each main chain
iteration, with variables along columns and iterations along rows.
\item \code{statistics}: a dataframe containing transition statistics for each main
chain iteration, with statistics along columns and iterations along rows.
}
}
\description{
Sample a Markov chain using Metropolis-Hastings kernel with given proposal
and target distributions, optionally adapting proposal parameters in warm-up
stage.
}
\examples{
target_distribution <- list(
  log_density = function(x) -sum(x^2) / 2,
  grad_log_density = function(x) -x
)
proposal <- barker_proposal(target_distribution, scale = 1.)
n_warm_up_iteration <- 1000
n_main_iteration <- 100
initial_state <- chain_state(rnorm(2))
results <- sample_chain(
  target_distribution,
  proposal,
  initial_state,
  n_warm_up_iteration,
  n_main_iteration
)
}
