% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptation.R
\name{scale_adapter}
\alias{scale_adapter}
\title{Create object to adapt proposal scale to coerce average acceptance rate.}
\usage{
scale_adapter(
  proposal,
  initial_scale = NULL,
  target_accept_prob = NULL,
  kappa = 0.6
)
}
\arguments{
\item{proposal}{Proposal object to adapt. Must define an \code{update} function
which accepts a parameter \code{scale} for setting scale parameter of proposal.}

\item{initial_scale}{Initial value to use for scale parameter. If not set
explicitly a proposal and dimension dependent default will be used.}

\item{target_accept_prob}{Target value for average accept probability for
chain. If not set a proposal dependent default will be used.}

\item{kappa}{Decay rate exponent in \verb{[0.5, 1]} for adaptation learning rate.}
}
\value{
List of functions with entries
\itemize{
\item \code{initialize}, a function for initializing adapter state at beginning of
chain,
\item \code{update} a function for updating adapter state and proposal parameters on
each chain iteration,
\item \code{finalize} a function for performing any final updates to adapter state and
proposal parameters on completion of chain sampling (may be \code{NULL} if
unused).
\item \code{state} a zero-argument function for accessing current values of adapter
state variables.
}
}
\description{
Create object to adapt proposal scale to coerce average acceptance rate.
}
\examples{
target_distribution <- list(
  log_density = function(x) -sum(x^2) / 2,
  grad_log_density = function(x) -x
)
proposal <- barker_proposal(target_distribution)
adapter <- scale_adapter(
  proposal,
  initial_scale = 1., target_accept_prob = 0.4
)
}
