% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridges.R
\name{target_distribution_from_stan_model}
\alias{target_distribution_from_stan_model}
\title{Construct target distribution from a BridgeStan \code{StanModel} object.}
\usage{
target_distribution_from_stan_model(model)
}
\arguments{
\item{model}{Stan model object to use for target (posterior) distribution.}
}
\value{
A list with entries
\itemize{
\item \code{log_density}: A function to evaluate log density function for target
distribution given current position vector.
\item \code{value_and_gradient_log_density}: A function to evaluate value and gradient
of log density function for target distribution given current position
vector, returning as a list with entries \code{value} and \code{gradient}.
}
}
\description{
Construct target distribution from a BridgeStan \code{StanModel} object.
}
\examples{
\dontshow{if (requireNamespace("bridgestan", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
model <- example_gaussian_stan_model()
target_distribution <- target_distribution_from_stan_model(model)
withr::with_seed(
  876287L, state <- chain_state(stats::rnorm(model$param_unc_num()))
)
state$log_density(target_distribution)
\dontshow{\}) # examplesIf}
}
