% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptation.R
\name{variance_adapter}
\alias{variance_adapter}
\title{Create object to adapt proposal with per dimension scales based on estimates
of target distribution variances.}
\usage{
variance_adapter(proposal, kappa = 0.6)
}
\arguments{
\item{proposal}{Proposal object to adapt. Must define an \code{update} function
which accepts a parameter \code{shape} for setting shape parameter of proposal.}

\item{kappa}{Decay rate exponent in \verb{[0.5, 1]} for adaptation learning rate.}
}
\value{
List of functions with entries \code{initialize}, a function for
initializing adapter state at beginning of chain, \code{update} a function for
updating adapter state and proposal parameters on each chain iteration,
and \code{finalize} a function for performing any final updates to adapter
state and proposal parameters on completion of chain sampling.
}
\description{
Create object to adapt proposal with per dimension scales based on estimates
of target distribution variances.
}
